Fixed explanations of speed
Fixed explanations of speed according to the change in the code.  It doesn't make sense to store the values in different units because the conversion formula is very simple.
Speed is changed from a tuple of values of different units to a single floating point value (in knots).
Use the helper function (speed_string) to convert it to kph or mph.